// src/models/EmployeeReq.ts
export class EmployeeReq {
  firstName: string;
  lastName: string;
  middleName: string;
  mobileNumbers: string;
  email: string;
  primaryAddressIndex: number;
  gender: string;
  roles: string;
  dateOfBirth: Date;
  userType: string;
  dateOfExist: Date;
  dateOfJoin: Date;
  salary: number;
  isCurrent: boolean;
  passcode: string;
  userName: string;
}